[carwow] Set pipeline number as integer
Fork of eddiewebb/queue v3 pre-release.

The pipeline values are not passing thru the orb, so they need to be
declared explicitly. Only the pipeline number is really neeed, but it
needs to be of type integer as CircleCI can't cast pipeline values.

https://discuss.circleci.com/t/pipeline-variables-not-valid-in-orbs/37221
https://discuss.circleci.com/t/unable-to-force-pipeline-number-to-be-type-string/40201

Also, set it up so we can publish a forked version within carwow namespace.