Raise jobs API limit to 100, and better handle missing "front_of_queue_pipeline_number"
If front_of_queue_pipeline_number is missing, it means there are more
jobs runnning than the ones the API returned.

I guess the behaviour here could be configurable, but I am choosing to
wait for the number of jobs to reduce, since otherwise we can't know if
we are front of the queue or not.