Cancel build if pipeline number is newer
If pipeline number is newer, that means we have reach the API limit of
running jobs, and newer pipeline already got front of the line, so we
can cancel the current build as it is already out of date.

In our use case this means we would try to deploy an older commit which
would then get rejected since a newer one is already deployed.

Also, reduce max wait time by default, we never want to wait for that
long. Plus, drop "dont-quit" and "force-cancel-previous", we don't use
them and not implemented.